refactor: 타입 생성
