+++ src/types/modal.d.ts
@@ -1,0 +1,6 @@
+export interface selectBookProps {
+  isbn: string;
+  title: string;
+  thumbnail: string;
+  author: string;
+}
